API needs no lock
