+++ bookwyrm/views/relationships.py
@@ -9,7 +9,8 @@ from django.views.decorators.vary import vary_on_headers
 
 from bookwyrm.activitypub import ActivitypubResponse
 from bookwyrm.settings import PAGE_LENGTH
-from .helpers import get_user_from_username, is_api_request, PrivateProfileMixin
+from bookwyrm.views.helpers import is_api_request
+from bookwyrm.views.mixins import PrivateProfileMixin
 
 
 class Relationships(PrivateProfileMixin, View):
